+++ src/cli/outputs.go
@@ -3,7 +3,7 @@ package cli
 import (
 	"fmt"
 
-	gcli "github.com/urfave/cli"
+	gcli "github.com/spf13/cobra"
 
 	"github.com/skycoin/skycoin/src/api/webrpc"
 	"github.com/skycoin/skycoin/src/cipher"
@@ -11,52 +11,45 @@ import (
 	"github.com/skycoin/skycoin/src/wallet"
 )
 
-func walletOutputsCmd(cfg Config) gcli.Command {
-	name := "walletOutputs"
-	return gcli.Command{
-		Name:      name,
-		Usage:     "Display outputs of specific wallet",
-		ArgsUsage: "[wallet file]",
-		Description: fmt.Sprintf(`Display outputs of specific wallet, the default
-		wallet (%s) will be
-		used if no wallet was specified, use ENV 'WALLET_NAME'
-		to update default wallet file name, and 'WALLET_DIR' to update
-		the default wallet directory`, cfg.FullWalletPath()),
-		OnUsageError: onCommandUsageError(name),
-		Action:       getWalletOutputsCmd,
+func walletOutputsCmd() *gcli.Command {
+	return &gcli.Command{
+		Short: "Display outputs of specific wallet",
+		Use:   "walletOutputs [wallet file]",
+		Long: fmt.Sprintf(`Display outputs of specific wallet, the default wallet (%s) will be
+    used if no wallet was specified, use ENV 'WALLET_NAME'
+    to update default wallet file name, and 'WALLET_DIR' to update
+    the default wallet directory`, cliConfig.FullWalletPath()),
+		DisableFlagsInUseLine: true,
+		SilenceUsage:          true,
+		RunE:                  getWalletOutputsCmd,
 	}
 }
 
-func addressOutputsCmd() gcli.Command {
-	name := "addressOutputs"
-	return gcli.Command{
-		Name:      name,
-		Usage:     "Display outputs of specific addresses",
-		ArgsUsage: "[address list]",
-		Description: `Display outputs of specific addresses, join multiple addresses with space,
-        example: addressOutputs $addr1 $addr2 $addr3`,
-		OnUsageError: onCommandUsageError(name),
-		Action:       getAddressOutputsCmd,
+func addressOutputsCmd() *gcli.Command {
+	return &gcli.Command{
+		Short: "Display outputs of specific addresses",
+		Use:   "addressOutputs [address list]",
+		Long: `Display outputs of specific addresses, join multiple addresses with space,
+    example: addressOutputs $addr1 $addr2 $addr3`,
+		Args:                  gcli.MinimumNArgs(1),
+		DisableFlagsInUseLine: true,
+		SilenceUsage:          true,
+		RunE:                  getAddressOutputsCmd,
 	}
-
 }
 
-func getWalletOutputsCmd(c *gcli.Context) error {
-	cfg := ConfigFromContext(c)
-	client := APIClientFromContext(c)
-
-	w := ""
-	if c.NArg() > 0 {
-		w = c.Args().First()
+func getWalletOutputsCmd(c *gcli.Command, args []string) error {
+	var wltPath string
+	if len(args) == 1 {
+		wltPath = args[0]
 	}
-
 	var err error
-	w, err = resolveWalletPath(cfg, w)
+	w, err := resolveWalletPath(cliConfig, wltPath)
 	if err != nil {
 		return err
 	}
 
-	outputs, err := GetWalletOutputsFromFile(client, w)
+	outputs, err := GetWalletOutputsFromFile(apiClient, w)
 	if err != nil {
 		return err
 	}
@@ -66,19 +59,18 @@ func getWalletOutputsCmd(c *gcli.Context) error {
 	})
 }
 
-func getAddressOutputsCmd(c *gcli.Context) error {
-	client := APIClientFromContext(c)
+func getAddressOutputsCmd(c *gcli.Command, args []string) error {
+	addrs := make([]string, len(args))
 
-	addrs := make([]string, c.NArg())
 	var err error
-	for i := 0; i < c.NArg(); i++ {
-		addrs[i] = c.Args().Get(i)
+	for i := 0; i < len(args); i++ {
+		addrs[i] = args[i]
 		if _, err = cipher.DecodeBase58Address(addrs[i]); err != nil {
 			return fmt.Errorf("invalid address: %v, err: %v", addrs[i], err)
 		}
 	}
 
-	outputs, err := client.OutputsForAddresses(addrs)
+	outputs, err := apiClient.OutputsForAddresses(addrs)
 	if err != nil {
 		return err
 	}